Refactor RequestResponseData to use public properties and set them in the constructor. Move interface to errors file.
