+++ src/apis/platformvm/api.ts
@@ -6,7 +6,8 @@ import { Buffer } from "buffer/"
 import BN from "bn.js"
 import AvalancheCore from "../../avalanche"
 import { JRPCAPI } from "../../common/jrpcapi"
-import { RequestResponseData, ErrorResponseObject } from "../../common/apibase"
+import { RequestResponseData } from "../../common/apibase"
+import { ErrorResponseObject } from "../../utils/errors"
 import BinTools from "../../utils/bintools"
 import { KeyChain } from "./keychain"
 import { Defaults, PlatformChainID, ONEAVAX } from "../../utils/constants"
