+++ tests/apis/auth/api.test.ts
@@ -2,7 +2,7 @@ import mockAxios from "jest-mock-axios"
 import { HttpResponse } from "jest-mock-axios/dist/lib/mock-axios-types"
 import { Avalanche } from "src"
 import { AuthAPI } from "../../../src/apis/auth/api"
-import { ErrorResponseObject } from "../../../src/common"
+import { ErrorResponseObject } from "../../../src/utils/errors"
 
 describe("Auth", (): void => {
   const ip: string = "127.0.0.1"
