+++ tests/apis/platformvm/api.test.ts
@@ -36,10 +36,8 @@ import {
   SerializedType
 } from "../../../src/utils/serialization"
 import { AddValidatorTx } from "../../../src/apis/platformvm/validationtx"
-import {
-  GetRewardUTXOsResponse,
-  ErrorResponseObject
-} from "../../../src/common"
+import { GetRewardUTXOsResponse } from "../../../src/common"
+import { ErrorResponseObject } from "../../../src/utils/errors"
 import { HttpResponse } from "jest-mock-axios/dist/lib/mock-axios-types"
 
 /**
